+++ modules/data/__init__.py
@@ -1,0 +1,1 @@
+"""Data utilities package."""
