+++ scripts/note_cli.py
@@ -1,4 +1,4 @@
-from modules.note_manager import run_note_manager
+from modules.management.note_manager import run_note_manager
 
 if __name__ == "__main__":
     run_note_manager()
